streamlit_poc/requirements.txt
streamlit>=1.31.0
langgraph>=0.0.20
langchain-core>=0.1.23
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.41.0
lxml>=4.9.0
numpy<2.0.0
